Session 11: 8 library fixes — @objc optional gate, SPI enum cases, autoclosure forwarding, CF opaque pointer dispatch, ObjC enum raw value types
Validation: 50/56 Swift wrapper targets passing (up from 42/56).
Libraries fixed: FSPagerView, Parchment, StripePaymentsUI, StripeConnect,
StripePaymentSheet, Alamofire, Kingfisher, StripePayments.

EC-20 (new): @objc optional method/property gate in WitnessDispatchEmitter
and EveryProtocolEmitter — skip dispatch for optional ObjC protocol members.
EC-10: @_spi enum case filtering in CaseByIndex wrapper and C# emission.
EC-12: @autoclosure forwarding with () in all 3 closure adapter call sites.
EC-13: UIKit/AVFoundation XML rawValueType Int64→Int, UInt64→UInt; enum
underlying type mapping Int→long, UInt→ulong for platform-width integers.
EC-4: UIFont.Weight, PHPickerResult added to AppleFrameworkRegistry ValueTypes.
EC-3: NSObjectProtocol detection via GenericSignature; static Self-typed
member gate in EveryProtocol conformance.
CF opaque pointer gate in WitnessDispatchEmitter.IsTypeBlittable — rejects
non-Swift-module types projected as System.IntPtr.
String return type annotation for overload disambiguation.
Actor isolation nested-type fallback for methods and properties.

+++ src/Swift.Bindings/tests/UnitTests/EmitterTests/EnumHandlerTests.cs
@@ -932,12 +932,12 @@ public class EnumHandlerTests
     [Fact]
     public void GetSwiftAbiMetadataType_Int_DiffersFromGetCSharpEnumUnderlyingType()
     {
-        // GetCSharpEnumUnderlyingType maps Int → "int" (4-byte, for C# enum declarations)
+        // GetCSharpEnumUnderlyingType maps Int → "long" (8-byte, for C# enum declarations matching 64-bit ABI)
         // GetSwiftAbiMetadataType maps Int → "nint" (pointer-sized, for Swift ABI metadata)
         var enumUnderlying = EnumHandler.GetCSharpEnumUnderlyingType("Int");
         var abiMetadata = EnumHandler.GetSwiftAbiMetadataType("Int");
 
-        Assert.Equal("int", enumUnderlying);
+        Assert.Equal("long", enumUnderlying);
         Assert.Equal("nint", abiMetadata);
         Assert.NotEqual(enumUnderlying, abiMetadata);
     }
